Some improvement on typstmathinput, estimate height instead of using left..right and allow more Unicode characters

+++ typstmathinput.py
@@ -4,6 +4,7 @@ from pathlib import Path
 import subprocess
 import shutil
 import hashlib
+import re
 
 tmpdir=Path(tempfile.gettempdir())/".typstmathinput-tmp"
 tmpdir.mkdir(exist_ok=True)
@@ -22,6 +23,8 @@ def typstmathinput_typst_formula_to_tex(s: str, template_hash: bytes)->str:
 """)
 	try: subprocess.run(["typst", n, n.with_suffix(".pdf")], cwd=tmpdir, check=True)
 	finally: n.unlink(missing_ok=True)
+	if not n.with_suffix(".pdf").is_file():
+		raise ValueError(f"Formula ${s}$ is invalid!")
 	try: subprocess.run(["pdftotext", n.with_suffix(".pdf")], cwd=tmpdir, check=True)
 	finally: n.with_suffix(".pdf").unlink(missing_ok=True)
 	try: result: str=n.with_suffix(".txt").read_text().replace("\n", "").replace("\x0c", "")
@@ -31,6 +34,15 @@ def typstmathinput_typst_formula_to_tex(s: str, template_hash: bytes)->str:
 def typstmathinput_printdollar_py()->str:
 	from pythonimmediate import get_arg_str, print_TeX
 	s = get_arg_str()
+
+	# preprocess it before passing to Typst
+	superscript = dict(zip("⁰¹²³⁴⁵⁶⁷⁸⁹⁺⁻⁼", "0123456789+-="))
+	subscript = dict(zip("₀₁₂₃₄₅₆₇₈₉₊₋₌", "0123456789+-="))
+	s = re.sub('[' + "".join(superscript) + ']+', lambda x: "^(" + x[0].translate(str.maketrans(superscript)) + ")" , s)
+	s = re.sub('[' + "".join(subscript) + ']+', lambda x: "_(" + x[0].translate(str.maketrans(subscript)) + ")" , s)
+	s = re.sub(r'√\s*\(', 'sqrt(', s)
+	s = re.sub(r'√\s*(\d+|\w+)', r'sqrt(\1)', s)
+
 	print_TeX(typstmathinput_typst_formula_to_tex(
 		s,
 		hashlib.sha256(template_file.read_bytes()).digest()
